Custom Scripts for w10/Clickables/interactive power options app/requirements.txt
altgraph==0.17.4
MouseInfo==0.1.3
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pystray==0.19.5
pytweening==1.2.0
pywin32-ctypes==0.2.3
setuptools==75.6.0
six==1.17.0
tkinter-tooltip==3.1.2
tooltip==1.0.0
